* Makefile.W32: Changed to dynamic linking  * config.h: Removed one warning "W4251". We need to find out how to avoid it. This helps seeing    other warnings. Remove this fix later.  * SipMessageTransport: Fix so that minisip can be compiled with debugging in MSVC (add function    to turn on/off of packet output). I had trouble declaring external variables that exist in a .dll. This   removes the only place we have such a hack.  * Many places: Removed warnings (the typical case was where an int was set to a size_t return type).  * Included config.h where it was not done (this is needed when using MSVC to get the LIBMSIP_API    definition correct).
git-svn-id: svn://svn.minisip.org/minisip/trunk@1336 63dae60f-c5d3-0310-92a3-fcbf00261c72

+++ libmsip/source/SipMIMEContent.cxx
@@ -22,6 +22,8 @@
  *          
 */
 
+#include<config.h>
+
 #include<libmsip/SipMIMEContent.h>
 #include<libmsip/SipMessage.h>
 #include<libmsip/SipMessageContentFactory.h>
@@ -46,7 +48,7 @@ SipMimeContent::SipMimeContent(std::string ContentType, std::string Message, std
 }
 
 SipMimeContent::SipMimeContent(std::string content, std::string ContentType) {
-	int index2;
+	size_t index2;
 	std::string cont;
 	this->uniqueboundry = "_Minisip";
 	if(ContentType.substr(0,9) == "multipart"){
@@ -63,10 +65,10 @@ SipMimeContent::SipMimeContent(std::string content, std::string ContentType) {
 		else
 			this->Message = "";
 		// Extract the bodyparts
-		int boundrysize = 2 + this->boundry.length();
+		size_t boundrysize = 2 + this->boundry.length();
 		// Find end of body
-		int endindex = content.rfind("--"+this->boundry+"--", content.length());
-		int index1 = index2;
+		size_t endindex = content.rfind("--"+this->boundry+"--", content.length());
+		size_t index1 = index2;
 		while (endindex != index1){
 			index1 = index1 + boundrysize + 2;
 			if (content.substr(index1,2) == "\r\n"){
